Update TextView to automatically resize text
Our TextView will now automatically resize text until it fits in the
text view's area.  If there is too much text, it will append an
ellipsis to the end of the string, therefore truncating it.

On my tablet, the IPv6 address displays properly, without bleeding into
the images of the devices.

Possible fix for issue #5. Needs more testing