Start to fix some issues with the AcceptThread
AcceptThread (Wifi) now knows how many active connections there are,
and how many there should be.  It doesn't stop accepting until it is
told to.  Instead, when someone tries to connect, we check to see if we
have too many connections already, and if so, just disconnect from the
newly connected player.