Rewrite connection screen
Issue #8 requests a rewrite of the connection screen

This rewrite clears up a lot of the confusion dealing with
disconnections. Now it uses the actual game instance to decide where
players should be added, and how many people are disconnected. In
theory this should allow multiple disconnections to happen at the same
time, and allow multiple people to be reconnecting at the same time.

Things left to do:
-Check to see how the Gameboard starts the ConnectActivity, so that we
can guarantee that we only get one instance of the ConnectActivity
starting at a time.
-Verify that if someone disconnects when the ConnectActivity is open
they are either dropped (if the game hasn't been started), or listed as
disconnected allowing another user to take their place (if the game has
been started)