Work on the game board layout some more
The name no longer needs to be padded with spaces, and the the rotation
stuff all looks pretty good. I have noticed a couple issues with some
of the bottoms of the cards still displaying after we no longer have
the card, so that should probably be looked at.

Also, add a counter for the number of cards that the user has, but are
not displayed on the tablet (Issue #10).