Fix force close on some picky Honeycomb devices
Some Honeycomb devices, including the emulator, are picky about the
themes. We were using Theme.DeviceDefault as the parent for a number of
our themes, which works on most devices. But due to the fact this
DeviceDefault was introduced in ICS, we should specifically define a
parent theme that _will_ be there on each SDK. So for the base, use the
regular Theme.  For Honeycomb, use Theme.Holo, and for ICS and up, use
Theme.DeviceDefault.

This should fix Issue #23.
Thanks to http://stackoverflow.com/a/9676533 for the idea.