Refactor refresh-player
The old version of refresh player used a JSONArray with specific items
at specific indices (so the discard card was at index 2 in the array,
and then index 3 through the end was the player's hand).  This isn't
the best solution, as coordinating indices isn't practical. This change
keeps backwards compatibility, while also adding support for REFRESHV2,
which uses a JSONObject as the base wrapper, with items stored in it by
a string key.